blockvault-redactor/requirements.txt
PyMuPDF>=1.24.0
python-docx>=1.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
spacy>=3.7.0
pikepdf>=9.0.0
Pillow>=10.0.0
presidio-analyzer>=2.2.351
